PMA-127 - MessagePostPolicy

+++ app/Http/Controllers/MessagePostController.php
@@ -9,7 +9,7 @@ use App\Http\Requests\MessagePost\UpdateRequest;
 use App\Http\Resources\MessagePostResource;
 use App\Http\Resources\MessagePostResourceCollection;
 use App\Repositories\Contracts\MessagePostRepository;
-use Illuminate\Http\Request;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class MessagePostController extends Controller
 {
@@ -32,9 +32,12 @@ class MessagePostController extends Controller
      *
      * @param IndexRequest $request
      * @return MessagePostResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [MessagePost::class, $request->get('messageId')]);
+
         $messagePosts = $this->messagePostRepository->findBy($request->all());
 
         return new MessagePostResourceCollection($messagePosts);
@@ -43,11 +46,14 @@ class MessagePostController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param StoreRequest  $request
      * @return MessagePostResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [MessagePost::class, $request->get('messageId')]);
+
         $messagePost = $this->messagePostRepository->save($request->all());
 
         return new MessagePostResource($messagePost);
@@ -58,9 +64,12 @@ class MessagePostController extends Controller
      *
      * @param MessagePost $messagePost
      * @return MessagePostResource
+     * @throws AuthorizationException
      */
     public function show(MessagePost $messagePost)
     {
+        $this->authorize('show', $messagePost);
+
         return new MessagePostResource($messagePost);
     }
 
@@ -70,9 +79,12 @@ class MessagePostController extends Controller
      * @param UpdateRequest $request
      * @param MessagePost $messagePost
      * @return MessagePostResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, MessagePost $messagePost)
     {
+        $this->authorize('update', $messagePost);
+
         $messagePost = $this->messagePostRepository->update($messagePost, $request->all());
 
         return new MessagePostResource($messagePost);
@@ -83,9 +95,12 @@ class MessagePostController extends Controller
      *
      * @param MessagePost $messagePost
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(MessagePost $messagePost)
     {
+        $this->authorize('destroy', $messagePost);
+
         $this->messagePostRepository->delete($messagePost);
 
         return response()->json(null, 204);
